services/shopperbot_service/requirements.txt
# FastAPI e servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client
httpx==0.25.2
requests==2.31.0

# Image processing
Pillow==10.1.0

# Database - PostgreSQL
psycopg2-binary==2.9.9

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Utils
python-dotenv==1.0.0

